Added Player Locomotion and Jumping
Character Controller being moved by Vector 3 manipulation via player input.

Jump only if CharacterController.isGrounded. If not, += gravity. 
Also while grounded, downward Vector3 is reset to a value close to zero. Otherwise it creates a bug that accumulates gravity if the player never jumps to reset it (to the JumpF value). For example, if the player keeps falling off of steps (without ever jumping), gravity will be normal for the first few steps, but gradually, the pull will appear as if the controller is teleporting to the ground because of the compounding strength of gravity because nothing is resetting it.